Remove icons, fonts, and debug_grid modules; extract _svg_to_png into svg_rasterize
Remove render/icons.py (IconManager, IconError, Iconify API),
render/fonts.py, and render/debug_grid.py — none are required for
dsui functionality.

Extract the _svg_to_png helper (the only piece dsui depends on)
into render/svg_rasterize.py with its own RasterizeError.

Update Deck to no longer carry an IconManager or debug_grid state,
simplify key_renderer and touch_renderer, drop httpx/respx deps,
and remove all related tests and documentation.
