+++ pyproject.toml
@@ -31,7 +31,6 @@ dependencies = [
     "pillow>=12.0.0",
     "streamdeck>=0.9.8",
     "cairosvg>=2.7.0",
-    "httpx>=0.27.0",
     "pyyaml>=6.0",
 ]
 
@@ -55,7 +54,6 @@ test = [
     "pytest>=8.0",
     "pytest-asyncio>=0.24",
     "pytest-cov>=6.0",
-    "respx>=0.22",
 ]
 
 [tool.pytest.ini_options]
