+++ src/deckboard/render/touch_renderer.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 from PIL import Image
 
-from .debug_grid import draw_touchscreen_grid
 from .key_renderer import _encode_jpeg
 from .metrics import (
     MARGIN_LEFT,
@@ -20,14 +19,12 @@ from .metrics import (
 
 def compose_touchstrip(
     cards: list[Image.Image | None],
-    debug_grid: bool = False,
     background: str = "black",
 ) -> bytes:
     """Compose 4 card images into a single touchscreen JPEG.
 
     Args:
         cards: Up to 4 card images (or ``None`` for blank slots).
-        debug_grid: When ``True``, overlay a 32x4 alignment grid.
         background: Fill colour for the 800x100 canvas (margins and gaps).
     """
     img = Image.new("RGB", (TOUCHSCREEN_WIDTH, TOUCHSCREEN_HEIGHT), background)
@@ -39,17 +36,11 @@ def compose_touchstrip(
             x = MARGIN_LEFT + index * (PANEL_WIDTH + PANEL_GAP)
             img.paste(card_image, (x, MARGIN_TOP))
 
-    if debug_grid:
-        img = draw_touchscreen_grid(img)
-
     return _encode_jpeg(img)
 
 
 def render_blank_touchscreen(
-    debug_grid: bool = False,
     background: str = "black",
 ) -> bytes:
     """Render a blank touch-strip image."""
-    return compose_touchstrip(
-        [None] * PANEL_COUNT, debug_grid=debug_grid, background=background
-    )
+    return compose_touchstrip([None] * PANEL_COUNT, background=background)
